feat(apperrors): refactor error handling by moving apperrors and interfaces package to internal/shared and updating imports
