+++ api/internal/modules/org/usecase.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/deveasyclick/openb2b/internal/model"
-	"github.com/deveasyclick/openb2b/pkg/apperrors"
+	"github.com/deveasyclick/openb2b/internal/shared/apperrors"
 	"github.com/deveasyclick/openb2b/pkg/interfaces"
 )
