+++ api/internal/shared/validator/validator.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/deveasyclick/openb2b/pkg/apperrors"
+	"github.com/deveasyclick/openb2b/internal/shared/apperrors"
 	"github.com/go-playground/validator/v10"
 )
