+++ api/pkg/interfaces/org.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/deveasyclick/openb2b/internal/model"
-	"github.com/deveasyclick/openb2b/pkg/apperrors"
+	"github.com/deveasyclick/openb2b/internal/shared/apperrors"
 )
 
 type OrgRepository interface {
