+++ api/pkg/interfaces/user.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/deveasyclick/openb2b/internal/model"
-	"github.com/deveasyclick/openb2b/pkg/apperrors"
+	"github.com/deveasyclick/openb2b/internal/shared/apperrors"
 )
 
 type UserService interface {
